[Feature/FP-24] 도메인 관련 테스트 코드 추가

+++ finpik-api/src/main/java/finpik/resolver/loanproduct/application/dto/RecommendedLoanProductDto.java
@@ -9,7 +9,7 @@ public class RecommendedLoanProductDto {
     private final String productName;
     private final Float minInterestRate;
     private final Float maxInterestRate;
-    private final Integer maxLoanLimitAmount;
+    private final Long maxLoanLimitAmount;
 
     public RecommendedLoanProductDto(RecommendedLoanProduct recommendedLoanProduct) {
         this.loanProductId = recommendedLoanProduct.getLoanProductId();
