+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/loanproduct/projection/RecommendedLoanProductProjection.java
@@ -7,6 +7,6 @@ public record RecommendedLoanProductProjection(
     String productName,
     Float minInterestRate,
     Float maxInterestRate,
-    Integer maxLoanLimitAmount
+    Long maxLoanLimitAmount
 ) {
 }
